Move reduce_configuration() and get_term_as_tuple() to their readers
Each had a single production caller in the reader that owns its naming
convention: get_term_as_tuple() reads a Hillier level name, and
reduce_configuration() normalises a Nahar core-state configuration for
comparison in get_naharphotoion_upperlevelids().

get_term_as_tuple() moves to readhillierdata, using that module's own
lchars. reduce_configuration() moves to readnahardata, taking
remove_bracketed_part() with it, since reduce_configuration() was its
only caller. get_parity_from_config() and interpret_configuration()
stay in __init__: both have consumers in more than one reader.

No behaviour change: the cmfgen, qub and kurucz checksums and the Nahar
handler's adata.txt are all unchanged.
