+++ artisatomic/__init__.py
@@ -839,100 +839,6 @@ reversedalphabets = "zyxwvutsrqponmlkjihgfedcbaZYXWVUTSRQPONMLKJIHGFEDCBA "
 lchars = "SPDFGHIKLMNOPQRSTUVWXYZ"
 
 
-# reads a Hillier level name and returns the term
-# tuple (twosplusone, l, parity)
-def get_term_as_tuple(config: str) -> tuple[int, int, int]:
-    config = config.split("[", maxsplit=1)[0]
-
-    if "{" in config and "}" in config:  # JJ coupling, no L and S
-        if config[-1] == "e":
-            return (-1, -1, 0)
-
-        if config[-1] == "o":
-            return (-1, -1, 1)
-
-        print(f"WARNING: Can't read parity from JJ coupling state '{config}'")
-        return (-1, -1, -1)
-
-    lposition = -1
-    l = -1
-    for charpos, char in reversed(list(enumerate(config))):
-        if char in lchars:
-            lposition = charpos
-            l = lchars.index(char)
-            break
-    if lposition < 0:
-        if config[-1] == "e":
-            return (-1, -1, 0)
-        if config[-1] == "o":
-            return (-1, -1, 1)
-        return (-1, -1, -1)
-    try:
-        twosplusone = int(config[lposition - 1])  # could this be two digits long?
-        if lposition + 1 > len(config) - 1:
-            # No 'e'/'o' suffix to read the parity from. CMFGEN writes its merged high-l levels
-            # this way, with the orbital letter repeated as the term symbol (2s2_13w_2W,
-            # 2s2_2p3(4So)5z_5Z), so the term letter is a merge marker rather than a real L.
-            # Sum l over the occupied orbitals instead of assuming the term is even.
-            parity = get_parity_from_config(config)
-        elif config[lposition + 1] == "o":
-            parity = 1
-        elif config[lposition + 1] == "e":
-            parity = 0
-        elif config[lposition + 2] == "o":
-            parity = 1
-        elif config[lposition + 2] == "e":
-            parity = 0
-        else:
-            twosplusone = -1
-            l = -1
-            parity = -1
-    #        sys.exit()
-    except (IndexError, ValueError):
-        twosplusone = -1
-        l = -1
-        parity = -1
-    return (twosplusone, l, parity)
-
-
-# e.g. convert "3d64s  (6D ) 8p  j5Fo" to "3d64s8p_5Fo",
-# similar to Hillier style "3d6(5D)4s8p_5Fo" but without the parent term
-# (and mysterious letter before the term if present)
-def reduce_configuration(instr: str) -> str:
-    if instr == "-1":
-        return "-1"
-    instr = instr.split("[", maxsplit=1)[0]  # remove trailing bracketed J value
-
-    if instr[-1] not in ["o", "e"]:
-        instr = instr + "e"  # last character being S,P,D, etc means even
-    if str.isdigit(instr[-2]):  # J value is in the term, so remove it
-        instr = instr[:-2] + instr[-1]
-
-    outstr = remove_bracketed_part(instr)
-    outstr += "_"
-    outstr += instr[-3:-1]
-    outstr += "o" if instr[-1] == "o" else "e"
-    return outstr
-
-
-def remove_bracketed_part(instr: str) -> str:
-    """Operates on a string by removing anything between parentheses (including the parentheses)
-    e.g. remove_bracketed_part('AB(CD)EF') = 'ABEF'.
-    """
-    outstr = ""
-    in_brackets = False
-    for char in instr[:-4]:
-        if char in (" ", "_"):
-            continue
-        if char == "(":
-            in_brackets = True
-        elif char == ")":
-            in_brackets = False
-        elif not in_brackets:
-            outstr += char
-    return outstr
-
-
 def interpret_configuration(instr_orig: str) -> tuple[list[str], int, int, int, int]:
     max_n = 20  # maximum possible principle quantum number n
     instr = instr_orig
